updated spif header file.
