+++ test_code/spif_code/spif.h
@@ -1,21 +1,88 @@
 //************************************************//
 //*                                              *//
-//* functions to interact with the spif          *//
+//* functions to interact with spif              *//
 //*                                              *//
-//* lap - 03/08/2021                             *//
+//* lap - 21/06/2021                             *//
 //*                                              *//
 //************************************************//
 
 #ifndef __SPIF_IF_H__
 #define __SPIF_IF_H__
 
-#include <fcntl.h>
-#include <unistd.h>
-#include <string.h>
-#include <errno.h>
-
 #include <sys/mman.h>
-#include <sys/ioctl.h>
+#include <fcntl.h>
+
+
+//---------------------------------------------------------------
+// FPGA selection
+//NOTE: un-comment the required FPGA #define
+//---------------------------------------------------------------
+#define TARGET_XC7Z015   // Zynq7 on TE0715 board
+//#define TARGET_XCZU9EG   // Zynq Ultrascale+ on zcu102 board
+//---------------------------------------------------------------
+
+#define HW_NUM_PIPES      2
+
+
+// addresses of memory-mapped FPGA interfaces 
+
+#ifdef TARGET_XC7Z015
+
+// ---------------------------------
+// AXI_HP0 PL -> DDR interface
+
+// reserved DDR range for DMA buffers 
+#define DDR_RES_MEM_ADDR  0x3fffc000
+#define DDR_RES_MEM_SIZE  0x4000
+#define PIPE_MEM_SIZE     (DDR_RES_MEM_SIZE / 4)
+// ---------------------------------
+
+// ---------------------------------
+// AXI_GP0 PS -> PL interface
+
+// spif configuration registers
+#define APB_REGS_ADDR     0x43c20000
+#define APB_REGS_SIZE     0x10000
+
+// DMA controller configuration
+#define DMA_REGS_ADDR     0x40400000
+#define DMA_REGS_SIZE     0x10000
+// ---------------------------------
+
+#endif /* TARGET_XC7Z015 */
+
+#ifdef TARGET_XCZU9EG
+
+// ---------------------------------
+// AXI_HP0_FPD PL -> DDR interface
+
+// reserved DDR range for DMA buffers 
+#define DDR_RES_MEM_ADDR  0x7feff000
+#define DDR_RES_MEM_SIZE  0x1000
+// ---------------------------------
+
+// ---------------------------------
+// AXI_HPM0_FPD PS -> PL interface
+
+// DMA controller configuration
+#define DMA_REGS_ADDR     0xa0000000
+#define DMA_REGS_SIZE     0x10000
+// ---------------------------------
+
+#endif /* TARGET_XCZU9EG */
+
+// ---------------------------------
+// DMA controller registers
+#define DMACR             0   // control
+#define DMASR             1   // status
+#define DMASA             6   // source 
+#define DMALEN            10  // length
+
+// DMA controller commands/status
+#define DMA_RESET         0x04
+#define DMA_RUN           0x01
+#define DMA_IDLE_MSK      0x02
+// ---------------------------------
 
 
 //--------------------------------------------------------------------
@@ -48,87 +115,95 @@
 // ---------------------------------
 
 
-// ---------------------------------
-// spif operations
-// ---------------------------------
-// driver ioctl unreserved magic number (seq 0xf0 - 0xff)
-#define SPIF_IOCTL_TYPE      'i'
-#define SPIF_IOCTL_SEQ       0xf0
-#define SPIF_NUM_REGS        128
-
-// spif ioctl operation fields
-//NOTE: size field used to encode spif register!
-#define SPIF_OP_TYPE         (SPIF_IOCTL_TYPE << _IOC_TYPESHIFT)
-#define SPIF_OP_REQ_MSK      (((1 << (_IOC_NRBITS + _IOC_TYPEBITS)) - 1) << _IOC_NRSHIFT)
-#define SPIF_OP_REQ_SHIFT    _IOC_NRSHIFT
-#define SPIF_OP_REG_MSK      ((SPIF_NUM_REGS - 1) << _IOC_SIZESHIFT)
-#define SPIF_OP_REG_SHIFT    _IOC_SIZESHIFT
-#define SPIF_OP_RD           (_IOC_READ  << _IOC_DIRSHIFT)
-#define SPIF_OP_WR           (_IOC_WRITE << _IOC_DIRSHIFT)
-#define SPIF_OP_CMD          (_IOC_NONE  << _IOC_DIRSHIFT)
-
-// spif operations
-#define SPIF_OP_REQ(r)       (SPIF_OP_CMD | ((SPIF_OP_TYPE | (SPIF_IOCTL_SEQ + r))  << _IOC_NRSHIFT))
-#define SPIF_REG_RD          SPIF_OP_REQ(0)
-#define SPIF_REG_WR          SPIF_OP_REQ(1)
-#define SPIF_STATUS_RD       SPIF_OP_REQ(2)
-#define SPIF_TRANSFER        SPIF_OP_REQ(3)
-// ---------------------------------
-
-
-// ---------------------------------
-// spif file descriptor associated with /dev/spif
-static int spif_fd;
-static int dummy;    // convenient place holder
-// ---------------------------------
+static volatile uint * dma_registers;
+static volatile uint * apb_registers;
 
 
 //--------------------------------------------------------------------
-// set up access to spif through /dev/spif
+// setup spif DMA controller through the AXI memory-mapped interface
 //
-// returns -1 if error
+// returns address of DMA data buffer -- NULL if problems found
 //--------------------------------------------------------------------
-void * spif_setup (int buffer_size)
-{
-  // check requested buffer size
-  if (buffer_size > SPIF_BUF_MAX_SIZE) {
-    printf ("spif error: buffer size exceeds maximum\n");
+void * spif_setup (uint pipe, size_t buf_size) {
+  // check that requested pipe exists
+  if (pipe >= HW_NUM_PIPES) {
+    printf ("error: requested event pipe does not exist\n");
     return (NULL);
   }
 
-  // open spif device
-  spif_fd = open ("/dev/spif", O_RDWR | O_SYNC);
-  if (spif_fd == -1) {
-    switch (errno) {
-    case ENOENT:
-      printf ("spif error: no spif device\n");
-      break;
-    case EBUSY:
-      printf ("spif error: spif device busy\n");
-      break;
-    case ENODEV:
-      printf ("spif error: no connection to SpiNNaker\n");
-      break;
-    default:
-      printf ("spif error: [%s]\n", strerror (errno));
-    }
+  // check that requested size fits in the reserved memory space
+  if (buf_size > PIPE_MEM_SIZE) {
+    printf ("error: requested buffer size exceeds limit\n");
     return (NULL);
   }
 
-  // map spif memory into user-space buffer 
-  void * buffer = mmap (
-    NULL, buffer_size, PROT_READ | PROT_WRITE, MAP_SHARED,
-    spif_fd, 0
+  //NOTE: make sure that mapped memory is *not* cached!
+  int fd = open ("/dev/mem", O_RDWR | O_SYNC);
+  if (fd < 1) {
+    printf ("error: unable to access DMA memory space\n");
+    return (NULL);
+  }
+
+  // map reserved DDR memory to process address space
+  void * rsvd_mem = mmap (
+    NULL, DDR_RES_MEM_SIZE, PROT_READ | PROT_WRITE,
+    MAP_SHARED, fd, DDR_RES_MEM_ADDR
     );
 
-  if (buffer == MAP_FAILED) {
-    printf ("spif error: unable to map spif buffer [%s]\n", strerror (errno));
-    close (spif_fd);
+  if (rsvd_mem == MAP_FAILED) {
+    printf ("error: unable to access reserved memory space\n");
     return (NULL);
   }
 
-  return (buffer);
+  // map APB (configuration) registers to process address space
+  apb_registers = (uint *) mmap (
+    NULL, APB_REGS_SIZE, PROT_READ | PROT_WRITE,
+    MAP_SHARED, fd, APB_REGS_ADDR
+    );
+
+  // map DMA (configuration) registers to process address space
+  off_t dma_addr = DMA_REGS_ADDR + (pipe * DMA_REGS_SIZE);
+  dma_registers = (uint *) mmap (
+    NULL, DMA_REGS_SIZE, PROT_READ | PROT_WRITE,
+    MAP_SHARED, fd, dma_addr
+    );
+
+  // close /dev/mem and drop root privileges
+  close (fd);
+  if (setuid (getuid ()) < 0) {
+    (void) munmap (rsvd_mem, DDR_RES_MEM_SIZE);
+    printf ("error: unable to access DMA memory space\n");
+    return (NULL);
+  }
+
+  // check that everything went well
+  if ((apb_registers == MAP_FAILED) || (dma_registers == MAP_FAILED)) {
+    (void) munmap (rsvd_mem, DDR_RES_MEM_SIZE);
+    printf ("error: unable to access DMA memory space\n");
+    return (NULL);
+  }
+
+  // locate dma buffer according to pipe
+  void * dma_buffer = rsvd_mem + (pipe * PIPE_MEM_SIZE);
+
+  // configure/initialise DMA controller for transmission
+  // reset DMA controller - reset interrupts
+  dma_registers[DMACR] = DMA_RESET;
+
+  // wait for reset to complete
+  while (dma_registers[DMACR] & DMA_RESET) {
+    continue;
+  }
+
+  // start DMA controller
+  dma_registers[DMACR] = DMA_RUN;
+
+  // write dma_buffer physical address to DMA controller
+  dma_registers[DMASA] = (uint) DDR_RES_MEM_ADDR + (uint) (pipe * PIPE_MEM_SIZE);
+
+  return (dma_buffer);
 }
+//--------------------------------------------------------------------
 
 
 //--------------------------------------------------------------------
@@ -138,7 +213,6 @@ void * spif_setup (int buffer_size)
 //--------------------------------------------------------------------
 void spif_close (void)
 {
-  close (spif_fd);
 }
 
 
@@ -149,15 +223,9 @@ void spif_close (void)
 //--------------------------------------------------------------------
 int spif_read_reg (unsigned int reg)
 {
-  // encode spif read register request
-  unsigned int req = (reg << 16) | SPIF_REG_RD;
-
-  // send request to spif and convey result
-  //NOTE: ioctl never fails with this request
-  (void) ioctl (spif_fd, req, (void *) &dummy);
-
-  return dummy;
+  return apb_registers[reg];
 }
+//--------------------------------------------------------------------
 
 
 //--------------------------------------------------------------------
@@ -165,59 +233,30 @@ int spif_read_reg (unsigned int reg)
 //--------------------------------------------------------------------
 void spif_write_reg (unsigned int reg, int val)
 {
-  // encode spif read register request
-  unsigned int req = (reg << 16) | SPIF_REG_WR;
-
-  // send request to spif
-  //NOTE: ioctl never fails with this request
-  (void) ioctl (spif_fd, req, (void *) (long) val);
+  apb_registers[reg] = val;
 }
-// ---------------------------------
+//--------------------------------------------------------------------
 
 
 //--------------------------------------------------------------------
-// read spif busy status
+// check status of spif DMA controller
 //
-// returns 0 if spif idle (no ongoing transfer)
+// returns 0 if DMA controller idle
 //--------------------------------------------------------------------
-int spif_busy (void)
-{
-  // send request to spif and convey result
-  //NOTE: ioctl never fails with this request
-  (void) ioctl (spif_fd, SPIF_STATUS_RD, (void *) &dummy);
-
-  return dummy;
+uint spif_busy (void) {
+  return (!(dma_registers[DMASR] & DMA_IDLE_MSK));
 }
+//--------------------------------------------------------------------
 
 
 //--------------------------------------------------------------------
-// start a transfer to SpiNNaker
-//
-// returns 0 if transfer succeeds
+// trigger a transfer to SpiNNaker
 //--------------------------------------------------------------------
-int spif_transfer (int length)
-{
-  // send request to spif and convey result
-  return (ioctl (spif_fd, SPIF_TRANSFER, (void *) (long) length));
+void spif_transfer (uint size) {
+  // write length of data batch (in bytes!)
+  dma_registers[DMALEN] = size;
 }
-
-
 //--------------------------------------------------------------------
-// spif service request
-//
-// returns 0 if request succeeds
-//--------------------------------------------------------------------
-int spif_req (unsigned int req, int * val)
-{
-  // send request to spif and convey result
-  int rc = ioctl (spif_fd, req, (void *) val);
-  if (rc == -1) {
-    printf ("spif error: spif request %d failed [%s]\n", req, strerror (errno));
-    return (-1);
-  }
-
-  return (rc);
-}
 
 
 #endif /* __SPIF_IF_H__ */
